Move scenery entry
