+++ src/openrct2-ui/interface/ViewportInteraction.cpp
@@ -33,6 +33,7 @@
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/object/BannerObject.h>
 #include <openrct2/object/FootpathItemObject.h>
+#include <openrct2/object/SmallSceneryObject.h>
 #include <openrct2/object/WallObject.h>
 #include <openrct2/ride/Ride.h>
 #include <openrct2/ride/RideConstruction.h>
