+++ src/openrct2-ui/windows/TileInspector.cpp
@@ -24,6 +24,7 @@
 #include <openrct2/object/FootpathObject.h>
 #include <openrct2/object/FootpathRailingsObject.h>
 #include <openrct2/object/FootpathSurfaceObject.h>
+#include <openrct2/object/SmallSceneryObject.h>
 #include <openrct2/object/TerrainEdgeObject.h>
 #include <openrct2/object/TerrainSurfaceObject.h>
 #include <openrct2/object/WallObject.h>
