+++ src/openrct2/interface/Viewport.cpp
@@ -22,6 +22,7 @@
 #include "../entity/Guest.h"
 #include "../entity/PatrolArea.h"
 #include "../entity/Staff.h"
+#include "../object/SmallSceneryObject.h"
 #include "../object/WallObject.h"
 #include "../paint/Paint.h"
 #include "../profiling/Profiling.h"
@@ -34,7 +35,7 @@
 #include "../util/Math.hpp"
 #include "../world/Climate.h"
 #include "../world/Map.h"
-#include "../world/SmallScenery.h"
+#include "../world/Scenery.h"
 #include "Colour.h"
 #include "Window.h"
 #include "Window_internal.h"
