+++ src/openrct2/world/MapGen.cpp
@@ -19,6 +19,7 @@
 #include "../localisation/StringIds.h"
 #include "../object/Object.h"
 #include "../object/ObjectManager.h"
+#include "../object/SmallSceneryObject.h"
 #include "../object/TerrainEdgeObject.h"
 #include "../object/TerrainSurfaceObject.h"
 #include "../platform/Platform.h"
